@@ -8,142 +8,102 @@
   },
   "license": "Apache-2.0",
   "private": true,
-  "workspaces": [
-    "capi"
-  ],
   "dependencies": {
-    "@algolia/autocomplete-js": "^1.5.3",
-    "@algolia/autocomplete-plugin-query-suggestions": "^1.5.6",
-    "@algolia/autocomplete-plugin-recent-searches": "^1.5.6",
-    "@algolia/autocomplete-shared": "^1.5.6",
-    "@algolia/autocomplete-theme-classic": "^1.6.1",
-    "@algolia/client-search": "^4.13.0",
-    "@aws-amplify/amplify-cli-core": "^4.2.3",
-    "@aws-amplify/ui-components": "latest",
-    "@aws-amplify/ui-react": "^5.0.6",
-    "@cloudscape-design/components": "^3.0.238",
-    "@cloudscape-design/global-styles": "^1.0.8",
-    "@emotion/react": "^11.1.5",
-    "@emotion/styled": "^11.3.0",
-    "@mdx-js/mdx": "^2.0.0",
-    "@octokit/rest": "^19.0.7",
-    "algoliasearch": "^4.12.1",
-    "array-flatten": "^3.0.0",
-    "aws-amplify": "^5.0.5",
-    "copy-to-clipboard": "^3.2.1",
-    "dotenv": "^16.0.0",
-    "emotion": "^10.0.23",
-    "eslint-config-next": "^13.2.4",
-    "extract-mdx-metadata": "^2.0.0",
-    "gray-matter": "4.0.3",
-    "html-entities": "^1.2.1",
-    "instantsearch.js": "^4.39.1",
-    "next": "^13.2.4",
+    "@aws-amplify/amplify-cli-core": "^4.2.10",
+    "@aws-amplify/ui-react": "^6.0.7",
+    "@docsearch/react": "3",
+    "aws-amplify": "^6.0.9",
+    "next": "^13.5.6",
     "next-image-export-optimizer": "^1.8.3",
     "next-transpile-modules": "^9.0.0",
     "parse-imports": "^1.1.0",
-    "prismjs": "^1.21.0",
-    "ramda": "^0.28.0",
+    "prism-react-renderer": "^2.1.0",
+    "prismjs": "^1.27.0",
     "react": "^18.2.0",
+    "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.7.1",
-    "react-syntax-highlighter": "^15.5.0",
-    "rehype": "^11.0.0",
-    "rehype-img-size": "^1.0.1",
-    "remark": "^14.0.2",
-    "remark-mdx": "^2.0.0",
-    "remark-mdx-searchable": "^0.1.3",
-    "theme-ui": "^0.7.5",
-    "unified": "^10.1.2",
-    "unist-util-visit": "^4.1.0"
+    "react-icons": "^4.7.1"
   },
   "devDependencies": {
-    "@cloudscape-design/jest-preset": "^2.0.0",
-    "@mdx-js/loader": "^1.6.22",
-    "@next/mdx": "^10.1.3",
-    "@stencil/core": "^1.17.0",
-    "@stencil/eslint-plugin": "^0.2.1",
-    "@stencil/router": "^1.0.1",
-    "@stencil/sass": "^1.3.1",
-    "@testing-library/jest-dom": "^5.16.2",
+    "@mdx-js/loader": "^2.3.0",
+    "@mdx-js/mdx": "^2.3.0",
+    "@mdx-js/react": "^2.3.0",
+    "@next/bundle-analyzer": "^13.5.3",
+    "@next/mdx": "^13.5.4",
+    "@testing-library/jest-dom": "^6.1.2",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/fs-extra": "^9.0.1",
     "@types/jest": "^26.0.19",
     "@types/node": "^12.12.9",
-    "@types/react": "^17.0.3",
+    "@types/react": "^18.0.0",
     "@types/url-parse": "^1.4.3",
-    "@typescript-eslint/eslint-plugin": "^2.7.0",
-    "@typescript-eslint/parser": "^2.7.0",
+    "@typescript-eslint/eslint-plugin": "^6.13.1",
     "axios": "^1.3.4",
+    "classnames": "^2.3.2",
+    "cross-env": "^7.0.3",
     "cspell": "^6.2.3",
-    "eslint": "^6.6.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.16.0",
+    "dotenv": "^16.0.0",
+    "eslint": "8.54.0",
+    "eslint-config-next": "14.0.3",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-next": "^0.0.0",
+    "eslint-plugin-prettier": "^5.0.1",
     "fs-extra": "^9.0.1",
     "git-jiggy": "1.1.1",
     "husky": "^8.0.1",
     "jest": "^26.6.3",
     "jest-cli": "^26.6.3",
-    "prettier": "^1.19.1",
+    "json5": "^2.2.3",
+    "lint-staged": "^14.0.0",
+    "next-bundle-analyzer": "^0.6.7",
+    "prettier": "^3.0.3",
     "puppeteer": "^20.8.2",
+    "rehype": "^11.0.0",
+    "rehype-img-size": "^1.0.1",
+    "rehype-mdx-code-props": "^2.0.0",
+    "remark": "^14.0.2",
+    "remark-gfm": "^3.0.0",
+    "remark-mdx": "^2.3.0",
+    "remark-mdx-searchable": "^0.1.3",
     "rollup-plugin-node-polyfills": "^0.2.1",
     "sass": "^1.54.8",
+    "serve": "^14.2.1",
     "tiny-glob": "0.2.9",
     "ts-jest": "^26.0.3",
     "ts-node": "^8.5.0",
-    "typescript": "^4.3.2",
-    "workbox-build": "4.3.1"
+    "typescript": "^4.9",
+    "unist-util-visit": "^4.1.0"
   },
   "resolutions": {
     "**/trim": "0.0.3",
+    "@babel/core": "^7.23.2",
+    "@babel/traverse": "^7.23.2",
     "loader-utils": "2.0.4",
     "minimatch": "3.1.2",
-    "json5": "2.2.2",
     "decode-uri-component": "0.2.1",
     "fast-xml-parser": "4.2.5",
-    "proxy-agent": "^6.3.0",
     "semver": "7.5.2",
     "tough-cookie": "4.1.3",
-    "aws-cdk-lib": "2.80.0"
-  },
-  "jest": {
-    "preset": "./preset.js",
-    "roots": [
-      "<rootDir>/src"
-    ],
-    "transform": {
-      "^.+\\.(ts|tsx|js|mjs)$": [
-        "babel-jest",
-        {
-          "presets": [
-            "next/babel"
-          ]
-        }
-      ]
-    },
-    "testPathIgnorePatterns": [
-      "capi",
-      ".next",
-      "client"
-    ],
-    "setupFilesAfterEnv": [
-      "<rootDir>/jest.setup.js"
-    ]
+    "aws-cdk-lib": "2.80.0",
+    "prismjs": "^1.27.0",
+    "@adobe/css-tools": "4.3.2",
+    "follow-redirects": "^1.15.4"
   },
   "scripts": {
     "clean": "rm -rf node_modules yarn.lock",
     "refresh": "yarn clean && yarn",
     "test": "jest",
-    "dev": "next dev",
-    "spellcheck": "cspell \"src/**/*.mdx\"",
+    "dev": "yarn prebuild && next dev",
+    "spellcheck": "cspell \"src/**/*.mdx\" --no-progress",
     "spellcheck-diff": "git diff --name-only --cached | awk \"/src.*\\.mdx/{print}\" | npx cspell --no-must-find-files --file-list stdin",
-    "build": "node tasks/generate-sitemap.mjs && next build && next export -o client/www/next-build && next-image-export-optimizer --exportFolderPath client/www/next-build",
+    "build": "node tasks/generate-sitemap.mjs && next build",
+    "build:release": "yarn build && next-image-export-optimizer --exportFolderPath client/www/next-build",
     "next-build": "next build",
     "next-start": "next start",
-    "amplify:submissionsLambda": "cd amplify/backend/function/submissionsLambda/src && yarn install && yarn build && yarn test",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "analyze": "ANALYZE=true yarn next-build",
+    "prebuild": "node src/directory/generateDirectory.mjs && node src/directory/generateFlatDirectory.mjs",
+    "lint": "next lint"
   }
 }
