@@ -1,9 +1,9 @@
-module.exports = {
-  ANDROID_VERSION: "2.11.1",
-  ANDROID_DEVPREVIEW: "1.36.5-dev-preview.0",
-  ANDROID_V1_VERSION: "1.38.7",
-  ANDROID_V1_GEO_VERSION: "1.0.1",
-  ANDROID_V1_KOTLIN_VERSION: "0.22.7",
-  ANDROID_SDK_VERSION: "2.72.0",
-  KOTLIN_SDK_VERSION: "0.21.1-beta"
+export const versions = {
+  ANDROID_VERSION: '2.14.10',
+  ANDROID_DEVPREVIEW: '1.36.5-dev-preview.0',
+  ANDROID_V1_VERSION: '1.38.8',
+  ANDROID_V1_GEO_VERSION: '1.0.1',
+  ANDROID_V1_KOTLIN_VERSION: '0.22.8',
+  ANDROID_SDK_VERSION: '2.73.0',
+  KOTLIN_SDK_VERSION: '1.0.44'
 };
