@@ -61,38 +61,14 @@ function generatePathMap(
     '/lib-v1/q/platform/ios': {
       page: '/lib-v1/q/platform/[platform]'
     },
-    '/ui/q/framework/react': {
-      page: '/ui/q/framework/[framework]'
+    '/lib-v1/q/platform/flutter': {
+      page: '/lib-v1/q/platform/[platform]'
     },
-    '/ui/q/framework/react-native': {
-      page: '/ui/q/framework/[framework]'
+    '/lib-v1/q/platform/js': {
+      page: '/lib-v1/q/platform/[platform]'
     },
-    '/ui/q/framework/angular': {
-      page: '/ui/q/framework/[framework]'
-    },
-    '/ui/q/framework/vue': {
-      page: '/ui/q/framework/[framework]'
-    },
-    '/ui/q/framework/flutter': {
-      page: '/ui/q/framework/[framework]'
-    },
-    '/ui/q/framework/next': {
-      page: '/ui/q/framework/[framework]'
-    },
-    '/ui-legacy/q/framework/react': {
-      page: '/ui-legacy/q/framework/[framework]'
-    },
-    '/ui-legacy/q/framework/react-native': {
-      page: '/ui-legacy/q/framework/[framework]'
-    },
-    '/ui-legacy/q/framework/angular': {
-      page: '/ui-legacy/q/framework/[framework]'
-    },
-    '/ui-legacy/q/framework/vue': {
-      page: '/ui-legacy/q/framework/[framework]'
-    },
-    '/ui-legacy/q/framework/next': {
-      page: '/ui-legacy/q/framework/[framework]'
+    '/lib-v1/q/platform/react-native': {
+      page: '/lib-v1/q/platform/[platform]'
     },
     '/sdk/q/platform/js': {
       page: '/sdk/q/platform/[platform]'
@@ -104,6 +80,9 @@ function generatePathMap(
       page: '/sdk/q/platform/[platform]'
     },
     '/sdk/q/platform/flutter': {
+      page: '/sdk/q/platform/[platform]'
+    },
+    '/sdk/q/platform/react-native': {
       page: '/sdk/q/platform/[platform]'
     },
     '/console': {
@@ -146,19 +125,6 @@ function generatePathMap(
         filterKind = 'platform';
       } else if (route.startsWith('/start')) {
         filterKind = 'integration';
-      }
-
-      if (filterKind !== '') {
-        const aOrAn = 'aeiou'.includes(filterKind[0]) ? 'an' : 'a';
-        pathMap[route] = {
-          page: '/ChooseFilterPage',
-          query: {
-            address: route,
-            directoryPath: '/ChooseFilterPage',
-            filterKind: filterKind,
-            message: `Choose ${aOrAn} ${filterKind}:`
-          }
-        };
       }
     }
 
@@ -229,17 +195,6 @@ function generatePathMap(
         page: `${route}/q/${routeType}/[${routeType}]`
       };
     });
-    const aOrAn = 'aeiou'.includes(routeType[0]) ? 'an' : 'a';
-    pathMap[route] = {
-      page: '/ChooseFilterPage',
-      query: {
-        address: route,
-        directoryPath: '/ChooseFilterPage',
-        filterKind: routeType,
-        filters: filters,
-        message: `Choose ${aOrAn} ${routeType}:`
-      }
-    };
   }
   return pathMap;
 }
