@@ -1,4 +1,3 @@
 const tsPreset = require('ts-jest/presets/js-with-babel/jest-preset');
-const cloudscapePreset = require('@cloudscape-design/jest-preset');
 
-module.exports = Object.assign(tsPreset, cloudscapePreset);
+module.exports = Object.assign(tsPreset);
