@@ -1,36 +1,31 @@
-import React from "react";
-import {ExternalLinkGraphic} from "./styles";
-import {trackExternalLink} from "../../utils/track";
+import React from 'react';
+import { trackExternalLink } from '../../utils/track';
 
 type ExternalLinkProps = {
+  children: React.ReactNode;
   graphic?: string;
   href: string;
   anchorTitle?: string;
+  icon?: boolean;
+  className?: string;
 };
 
 const ExternalLink: React.FC<ExternalLinkProps> = ({
   children,
-  graphic,
   href,
-  anchorTitle,
+  className
 }) => {
   return (
     <a
       href={href}
+      className={className}
       rel="noopener noreferrer"
       target="_blank"
-      title={anchorTitle}
-      onClick={(e) => {
+      onClick={() => {
         trackLink(href);
       }}
     >
       {children}
-      {graphic && (
-        <ExternalLinkGraphic
-          alt="External link"
-          src={`/assets/external-link-${graphic}.svg`}
-        />
-      )}
     </a>
   );
 };
