@@ -1,33 +1,17 @@
+import { Fragment } from 'react';
 import FilterChildren from '../FilterChildren';
-import { useLastUpdatedDatesContext } from '../LastUpdatedProvider';
-import { MdxFrontmatterType } from '../Page';
 
 export default function Fragments({ fragments }) {
-  const children = [];
-  let frontmatter: MdxFrontmatterType;
+  const children: React.ReactNode[] = [];
 
-  const { state, dispatch } = useLastUpdatedDatesContext();
+  let filterKey = '';
 
   for (const key in fragments) {
+    if (!filterKey) filterKey = key;
     const fragment = fragments[key]([]);
-    frontmatter = fragment.props.frontmatter;
 
-    if (frontmatter && frontmatter.lastUpdated) {
-      if (
-        state.files[key] === undefined ||
-        (state.files[key] &&
-          !state.files[key].includes(frontmatter.lastUpdated))
-      ) {
-        dispatch({
-          type: 'update',
-          key: key,
-          lastUpdated: frontmatter.lastUpdated
-        });
-      }
-    }
-
-    children.push(<div key={key}>{fragment}</div>);
+    children.push(<Fragment key={key}>{fragment}</Fragment>);
   }
 
-  return <FilterChildren>{children}</FilterChildren>;
+  return <FilterChildren filterKey={filterKey}>{children}</FilterChildren>;
 }
