@@ -1,53 +1,78 @@
-import { Flex, Text, Button, VisuallyHidden } from '@aws-amplify/ui-react';
+import {
+  Flex,
+  Text,
+  Button,
+  VisuallyHidden,
+  Badge
+} from '@aws-amplify/ui-react';
 import styles from '../GlobalNav.module.scss';
 import { AmplifyLogo, ChevronIcon } from './icons';
-
-export function AmplifyNavLink({ currentSite, isCollapsed, setIsCollapsed }) {
+import Link from 'next/link';
+export function AmplifyNavLink({
+  currentSite,
+  isCollapsed,
+  setIsCollapsed,
+  isGen2
+}) {
   const chevronRotation = isCollapsed ? '0' : '180';
   return (
     <Flex justifyContent={'space-between'} className={styles['logo-container']}>
-      <Flex
-        className={styles['desktop-only']}
-        gap="xs"
-        alignItems="center"
-        direction="row"
-      >
-        <AmplifyLogo />
-        <Text className={styles['dev-center-logo']}>
-          <span style={{ fontWeight: '400' }}>Amplify</span>{' '}
-          <span style={{ fontWeight: '300' }}>Dev Center</span>
-          <span className={styles['mobile-only']}>
-            {' '}
-            <ChevronIcon rotateDeg="270" /> {currentSite}
-          </span>
-        </Text>
-      </Flex>
+      {isGen2 ? (
+        <Link href="/gen2" className={styles['gen2-home']}>
+          <AmplifyLogo isGen2={isGen2} />
+          <Text as="span" className={styles['dev-center-logo']}>
+            Amplify code-first DX
+          </Text>
+          <Badge marginInlineStart="medium" className={styles['navbar-badge']}>
+            Preview
+          </Badge>
+        </Link>
+      ) : (
+        <>
+          <Flex
+            className={styles['desktop-only']}
+            gap="xs"
+            alignItems="center"
+            direction="row"
+          >
+            <AmplifyLogo />
+            <Text as="span" className={styles['dev-center-logo']}>
+              <span style={{ fontWeight: '400' }}>Amplify</span>{' '}
+              <span style={{ fontWeight: '300' }}>Dev Center</span>
+              <span className={styles['mobile-only']}>
+                {' '}
+                <ChevronIcon rotateDeg="270" /> {currentSite}
+              </span>
+            </Text>
+          </Flex>
 
-      <Flex
-        className={styles['mobile-only']}
-        gap="xs"
-        alignItems="center"
-        direction="row"
-        as="a"
-        href="/"
-        style={{ textDecoration: 'none', cursor: 'pointer' }}
-      >
-        <AmplifyLogo />
-        <Text className={styles['dev-center-logo']}>
-          <span style={{ fontWeight: '400' }}>Amplify</span>{' '}
-          <span style={{ fontWeight: '300' }}>Dev Center</span>
-          <span className={styles['mobile-only']}>
-            {' '}
-            <ChevronIcon rotateDeg="270" /> {currentSite}
-          </span>
-        </Text>
-      </Flex>
+          <Flex
+            className={`${styles['mobile-only']} ${styles['mobile-current-link']}`}
+            gap="xs"
+            alignItems="center"
+            direction="row"
+            as="a"
+            href="/"
+            style={{ textDecoration: 'none', cursor: 'pointer' }}
+          >
+            <AmplifyLogo />
+            <Text as="span" className={styles['dev-center-logo']}>
+              <span style={{ fontWeight: '400' }}>Amplify</span>{' '}
+              <span style={{ fontWeight: '300' }}>Dev Center</span>
+              <span className={styles['mobile-only']}>
+                {' '}
+                <ChevronIcon rotateDeg="270" /> {currentSite}
+              </span>
+            </Text>
+          </Flex>
+        </>
+      )}
 
       <Button
         aria-expanded={!isCollapsed}
         aria-controls="mobile-nav-links"
-        border="none"
-        backgroundColor="transparent"
+        variation="link"
+        size="large"
         className={`${styles['mobile-only']} ${styles['nav-expander']}`}
         onClick={() => {
           setIsCollapsed(!isCollapsed);
