@@ -6,7 +6,7 @@ import { NavMenuItem } from '../GlobalNav';
 
 export function NavMenuLink({
   navMenuItem,
-  currentMenuItem = '',
+  currentMenuItem = ''
 }: {
   navMenuItem: NavMenuItem;
   currentMenuItem: string;
@@ -17,11 +17,8 @@ export function NavMenuLink({
   const label: string = navMenuItem.label;
   const linkContent: JSX.Element =
     navMenuItem.type === 'EXTERNAL' ? (
-      <Text as="span" color="inherit">
-        <ExternalLink>{label}</ExternalLink>
-      </Text>
+      <ExternalLink>{label}</ExternalLink>
     ) : (
-
       <Text as="span" color="inherit" className={styles['icon-link']}>
         <IconLink iconType={navMenuItem.icon ? navMenuItem.icon : ''} />
       </Text>
@@ -29,18 +26,15 @@ export function NavMenuLink({
 
   if (navMenuItem.type === 'DEFAULT') {
     return (
-      <Link className={`${styles['nav-menu-item']}`} href={navMenuItem.url}>
-        <Text
-          as="span"
-          color="inherit"
-          className={
-            navMenuItem.label === currentMenuItem
-              ? styles['current-nav-menu-item']
-              : ''
-          }
-        >
-          {label}
-        </Text>
+      <Link
+        className={`${styles['nav-menu-item']} ${
+          navMenuItem.label === currentMenuItem
+            ? styles['current-nav-menu-item']
+            : ''
+        }`}
+        href={navMenuItem.url}
+      >
+        {label}
       </Link>
     );
   } else {
