@@ -1,75 +1,119 @@
-import {
-  NextPreviousContainerStyle,
-  NextPreviousLinkStyle,
-  NextPreviousTextStyle,
-} from "./styles";
-import InternalLink from "../InternalLink";
-import {
-  getChapterDirectory,
-  isProductRoot,
-} from "../../utils/getLocalDirectory";
-import type { DirectoryItem } from "../Menu/Directory";
+import Link from 'next/link';
+import { View, Flex } from '@aws-amplify/ui-react';
+import directory from '@/directory/directory.json';
+import { useRouter } from 'next/router';
+import { useCurrentPlatform } from '@/utils/useCurrentPlatform';
+import { IconChevron } from '@/components/Icons';
 
-type Directory = {
-  items: DirectoryItem[];
-}
+export const NextPrevious = () => {
+  const platform = useCurrentPlatform();
+  const router = useRouter();
+  const pathname = router.pathname;
 
-function Prev(item: DirectoryItem) {
-  return (
-    <InternalLink href={item.route}>
-      <NextPreviousLinkStyle>
-        <img src="/assets/arrow-left.svg" alt="" />
-        <NextPreviousTextStyle isPrevious={true}>
-          <span>previous</span>
-          <h4>{item.title}</h4>
-        </NextPreviousTextStyle>
-      </NextPreviousLinkStyle>
-    </InternalLink>
-  );
-}
+  const findDirectoryNodes = (
+    route,
+    dir = directory,
+    platform,
+    previous,
+    next
+  ) => {
+    const children = dir?.children?.filter((child) => {
+      return child?.platforms?.includes(platform);
+    });
+    if (dir.route === route) {
+      return { previous, next };
+    } else if (children && children.length) {
+      for (let i = 0; i < children.length; i++) {
+        const child = children[i];
+        const res = findDirectoryNodes(
+          route,
+          child,
+          platform,
+          children[i - 1],
+          children[i + 1]
+        );
+        if (res) return res;
+      }
+    }
+  };
 
-function Next(item: DirectoryItem) {
-  return (
-    <InternalLink href={item.route}>
-      <NextPreviousLinkStyle>
-        <NextPreviousTextStyle isPrevious={false}>
-          <span>next</span>
-          <h4>{item.title}</h4>
-        </NextPreviousTextStyle>
-        <img src="/assets/arrow-right.svg" alt="" />
-      </NextPreviousLinkStyle>
-    </InternalLink>
-  );
-}
-
-export default function NextPrevious({url, filterKey}) {
-  if (isProductRoot(url)) {
-    return <></>;
+  const { previous, next } =
+    findDirectoryNodes(pathname, directory, platform, null, null) || {};
+  let nextHref, prevHref;
+  if (next) {
+    nextHref = {
+      pathname: next.route,
+      query: {
+        platform
+      }
+    };
+  }
+  if (previous) {
+    prevHref = {
+      pathname: previous.route,
+      query: {
+        platform
+      }
+    };
   }
 
-  const chapterDirectory = getChapterDirectory(url) as Directory;
-  if (!chapterDirectory) {
-    return null;
-  }
-
-  let {items} = chapterDirectory;
-  items = items.filter((item) => {
-    if (!("filters" in item) || item.filters.includes(filterKey)) return true;
-    return false;
-  });
-  let itemIndex = -1;
-  for (let i = 0; i < items.length; ++i) {
-    if (url.startsWith(items[i].route)) itemIndex = i;
-  }
-
-  if (itemIndex === -1) {
-    return <></>;
-  }
+  const justifyContent =
+    next && previous ? 'space-between' : previous ? 'flex-start' : 'flex-end';
 
   return (
-    <NextPreviousContainerStyle>
-      {itemIndex !== 0 ? Prev(items[itemIndex - 1]) : <div />}
-      {itemIndex != items.length - 1 ? Next(items[itemIndex + 1]) : <div />}
-    </NextPreviousContainerStyle>
+    <Flex justifyContent={justifyContent} className="next-prev">
+      {previous && (
+        <Link href={prevHref}>
+          <Flex>
+            <IconChevron margin="auto" className="icon-rotate-90" />
+            <Flex direction="column" gap="0">
+              <View className="next-prev__label">PREVIOUS</View>
+              <View className="next-prev__title">{previous.title}</View>
+            </Flex>
+          </Flex>
+        </Link>
+      )}
+      {next && (
+        <Link href={nextHref}>
+          <Flex>
+            <Flex direction="column" gap="0">
+              <View className="next-prev__label">NEXT</View>
+              <View className="next-prev__title">{next.title}</View>
+            </Flex>
+            <IconChevron margin="auto" className="icon-rotate-90-reverse" />
+          </Flex>
+        </Link>
+      )}
+    </Flex>
   );
-}
+};
+
+export const NEXT_PREVIOUS_SECTIONS = [
+  '/start/getting-started/',
+  '/start/project-setup/',
+  '/build-ui/formbuilder/',
+  '/build-ui/uibuilder/',
+  '/build-a-backend/auth/',
+  '/build-a-backend/functions/',
+  '/build-a-backend/graphqlapi/',
+  '/build-a-backend/more-features/',
+  '/build-a-backend/push-notifications/',
+  '/build-a-backend/restapi/',
+  '/build-a-backend/storage/',
+  '/build-a-backend/utilities/',
+  '/tools/cli/auth/',
+  '/tools/cli/custom/',
+  '/tools/cli/migration/',
+  '/tools/cli/plugins/',
+  '/tools/cli/project/',
+  '/tools/cli/reference/',
+  '/tools/cli/restapi/',
+  '/tools/cli/start/',
+  '/tools/cli/teams/',
+  '/tools/cli/usage/',
+  '/tools/console/adminui/',
+  '/tools/console/auth/',
+  '/tools/console/data/',
+  '/tools/console/storage/',
+  '/tools/console/tutorial/'
+];
