@@ -1,4 +1,3 @@
-export const ALGOLIA_API_KEY = "24d37f059982b2f5ecf829afe93aed40";
-export const ALGOLIA_INDEX_NAME = "aws_amplify_new";
-export const UNINITIALIZED_SEARCH_INPUT_SELECTOR =
-  "#amplify-docs-search-input:not(.ds-input)";
+export const ALGOLIA_API_KEY = '096fb25a5abb771ed2ac4391e80673c4';
+export const ALGOLIA_INDEX_NAME = 'amplify';
+export const ALGOLIA_APP_ID = 'WUUEEESZ67';
