@@ -8,11 +8,11 @@ export const PRIVACY = 'https://aws.amazon.com/privacy/';
 export const MARKETING = 'https://aws.amazon.com/amplify/framework/';
 export const AWS_USER_GUIDE =
   'https://docs.aws.amazon.com/amplify/latest/userguide/welcome.html';
-export const IOS_REFERENCE = 'https://aws-amplify.github.io/amplify-swift/docs/';
+export const IOS_REFERENCE =
+  'https://aws-amplify.github.io/amplify-swift/docs/';
 export const ANDROID_REFERENCE =
   'https://aws-amplify.github.io/aws-sdk-android/docs/reference/';
 export const JS_REFERENCE = 'https://aws-amplify.github.io/amplify-js/api/';
 export const HOSTING_REFERENCE =
   'https://docs.aws.amazon.com/amplify/latest/userguide/getting-started.html';
-  export const FLUTTER_REFERENCE =
-  '/flutter-references';
+export const FLUTTER_REFERENCE = '/flutter-references';
