@@ -1,64 +1,58 @@
-import Layout from "../components/Layout";
-import ExternalLink from "../components/ExternalLink";
-import InternalLinkButton from "../components/InternalLinkButton";
-import styled from "@emotion/styled";
-import {MQDesktop} from "../components/media";
-import {useEffect, useState} from "react";
+import { useRouter } from 'next/router';
+import { useEffect, useState } from 'react';
+import { IconExternalLink } from '@/components/Icons';
+import { Button, Flex, Text, Heading } from '@aws-amplify/ui-react';
 
-export const Host = styled.div`
-  width: 100%;
-  height: calc(100vh - 10.75rem);
-  padding: 1rem;
+export const meta = {
+  title: '404',
+  description: ''
+};
 
-  ${MQDesktop} {
-    height: calc(100vh - 7.75rem);
-  }
-
-  width: 100%;
-  height: 100%;
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-
-  h1 {
-    margin-bottom: 0.5rem;
-  }
-
-  p {
-    margin-bottom: 1rem;
-    max-width: 32rem;
-  }
-`;
+export function getStaticProps() {
+  return {
+    props: {
+      meta,
+      showBreadcrumbs: false,
+      useCustomTitle: true
+    }
+  };
+}
 
 export default function Custom404() {
-  let [href, setHref] = useState("https://docs.amplify.aws");
+  const basePath = 'https://docs.amplify.aws';
+  const [href, setHref] = useState(basePath);
+  const path = useRouter().asPath;
   useEffect(() => {
-    setHref(window.location.href);
-  }, []);
+    setHref(basePath + path);
+  }, [path]);
   return (
-    <Layout>
-      <Host>
-        <h1>404</h1>
-        <p>
-          {`Apologies––we can't seem to find the page for which you're looking. If this is a mistake, please `}
-          <ExternalLink
-            href={`https://github.com/aws-amplify/docs/issues/new?title=[missing-page]&labels=v2&body=${encodeURI(
-              `**Page**: [\`${href}\`](${href})
+    <Flex className="four-oh-four">
+      <Heading level={1}>404</Heading>
+      <Text>
+        Apologies &#8212; we can&apos;t seem to find the page for which
+        you&apos;re looking. If this is a mistake, please file an issue and
+        we&apos;ll fix it ASAP.
+      </Text>
+      <Flex className="four-oh-four__cta">
+        <Button as="a" href="/" variation="primary">
+          Return to home page
+        </Button>
+        <Button
+          variation="link"
+          as="a"
+          rel="noopener noreferrer"
+          target="_blank"
+          gap="small"
+          href={`https://github.com/aws-amplify/docs/issues/new?title=[missing-page]&labels=v2&body=${encodeURI(
+            `**Page**: [\`${href}\`](${href})
 
-**Feedback**: <!-- your feedback here -->
-`,
-            )}`}
-          >
-            file an issue
-          </ExternalLink>
-          {` and we'll fix it ASAP.`}
-        </p>
-        <InternalLinkButton href="/">
-          Return to the landing page
-        </InternalLinkButton>
-      </Host>
-    </Layout>
+**Feedback**: <!-- your feedback here \ */}
+`
+          )}`}
+        >
+          File an issue <IconExternalLink />
+        </Button>
+      </Flex>
+    </Flex>
   );
 }
