@@ -1,101 +1,58 @@
 let configured = false;
-let firstPageOfVisit = true;
 let AWSCShortbread;
 let s;
 let AWSMA;
 
-if (typeof window !== "undefined") {
+if (
+  typeof window !== 'undefined' &&
+  typeof window.AWSCShortbread !== 'undefined'
+) {
   AWSCShortbread = window.AWSCShortbread;
   AWSMA = window.AWSMA;
   s = window.s;
-}
 
-if (!configured) {
-  if (typeof window !== "undefined") {
+  if (!configured) {
     AWSCShortbread({
-      domain: ".amplify.aws",
+      domain: '.amplify.aws'
     }).checkForCookieConsent();
-    if (typeof s != "undefined") s.trackExternalLinks = false;
+    if (typeof s != 'undefined') s.trackExternalLinks = false;
+    configured = true;
   }
-  configured = true;
 }
 
 export enum AnalyticsEventType {
-  PAGE_VISIT = "PAGE_VISIT",
-  INTERNAL_LINK_CLICK = "INTERNAL_LINK_CLICK",
-  EXTERNAL_LINK_CLICK = "EXTERNAL_LINK_CLICK",
-  PAGE_DATA_FETCH_EXCEPTION = "PAGE_DATA_FETCH_EXCEPTION",
+  PAGE_VISIT = 'PAGE_VISIT',
+  INTERNAL_LINK_CLICK = 'INTERNAL_LINK_CLICK',
+  EXTERNAL_LINK_CLICK = 'EXTERNAL_LINK_CLICK',
+  PAGE_DATA_FETCH_EXCEPTION = 'PAGE_DATA_FETCH_EXCEPTION'
 }
-
-interface AnalyticsEventPageVisit {
-  type: AnalyticsEventType.PAGE_VISIT;
-  attributes: {
-    url: string;
-    previousUrl: string;
-    referrer: string;
-  };
-}
-
-interface AnalyticsEventInternalLinkClick {
-  type: AnalyticsEventType.INTERNAL_LINK_CLICK;
-  attributes: {
-    from: string;
-    to: string;
-  };
-}
-
-interface AnalyticsEventExternalLinkClick {
-  type: AnalyticsEventType.EXTERNAL_LINK_CLICK;
-  attributes: {
-    from: string;
-    to: string;
-  };
-}
-
-interface AnalyticsEventPageDataFetchException {
-  type: AnalyticsEventType.PAGE_DATA_FETCH_EXCEPTION;
-  attributes: {
-    url: string;
-    exception: Error;
-  };
-}
-
-type AnalyticsEvent =
-  | AnalyticsEventPageVisit
-  | AnalyticsEventInternalLinkClick
-  | AnalyticsEventExternalLinkClick
-  | AnalyticsEventPageDataFetchException;
 
 export const trackPageVisit = (): void => {
-  if (
-    typeof window !== "undefined" &&
-    typeof s != "undefined" &&
-    !firstPageOfVisit
-  ) {
+  if (typeof window !== 'undefined' && typeof s != 'undefined') {
+    s.pageName = window.location.href;
     s.pageURL = window.location.href;
     s.t();
   }
-  firstPageOfVisit = false;
 };
 
 export const trackPageFetchException = (): void => {
-  if (typeof window !== "undefined" && typeof s != "undefined") {
+  if (typeof window !== 'undefined' && typeof s != 'undefined') {
     s.linkTrackVars =
-      "prop39,prop41,prop50,prop61,prop62,eVar39,eVar41,eVar50,eVar61,eVar62,eVar69";
-    s.tl(true, "o", "page fetch exception");
+      'prop39,prop41,prop50,prop61,prop62,eVar39,eVar41,eVar50,eVar61,eVar62,eVar69';
+    s.tl(true, 'o', 'page fetch exception');
   }
 };
 
 export const trackExternalLink = (hrefTo: string): void => {
-  if (typeof window !== "undefined" && typeof s != "undefined") {
+  if (typeof window !== 'undefined' && typeof s != 'undefined') {
     s.linkTrackVars =
-      "prop39,prop41,prop50,prop61,prop62,eVar39,eVar41,eVar50,eVar61,eVar62,eVar69";
-    s.tl(true, "e", hrefTo);
+      'prop39,prop41,prop50,prop61,prop62,eVar39,eVar41,eVar50,eVar61,eVar62,eVar69';
+    s.tl(true, 'e', hrefTo);
   }
 };
 
 export const setSearchQuery = (query: string): void => {
-  if (typeof window !== "undefined" && typeof s != "undefined") {
+  if (typeof window !== 'undefined' && typeof s != 'undefined') {
     s.eVar26 = query;
   }
 };
@@ -105,12 +62,12 @@ const triggerNoSearchResults = (query: string): void => {
   const resultCountBackup: number = parseInt(s.eVar27, 10);
 
   s.eVar26 = query;
-  s.eVar27 = "0"; // If it's the number 0, the variable won't be sent
+  s.eVar27 = '0'; // If it's the number 0, the variable won't be sent
   s.linkTrackVars =
-    "prop39,prop41,prop50,prop61,prop62,eVar26,eVar27,eVar39,eVar41,eVar50,eVar61,eVar62,eVar69,events";
-  s.linkTrackEvents = "event2";
-  s.events = "event2";
-  s.tl(true, "o", "internal search");
+    'prop39,prop41,prop50,prop61,prop62,eVar26,eVar27,eVar39,eVar41,eVar50,eVar61,eVar62,eVar69,events';
+  s.linkTrackEvents = 'event2';
+  s.events = 'event2';
+  s.tl(true, 'o', 'internal search');
 
   s.eVar26 = queryBackup;
   s.eVar27 = resultCountBackup.toString();
@@ -118,9 +75,9 @@ const triggerNoSearchResults = (query: string): void => {
 
 let noResultsTimeout: NodeJS.Timeout;
 export const setSearchResultCount = (resultCount: number): void => {
-  if (typeof window !== "undefined" && typeof s != "undefined") {
+  if (typeof window !== 'undefined' && typeof s != 'undefined') {
     s.eVar27 = resultCount.toString();
-    s.events = resultCount === 0 ? "event1" : "event2";
+    s.events = resultCount === 0 ? 'event1' : 'event2';
 
     if (resultCount === 0) {
       if (noResultsTimeout) {
@@ -128,25 +85,19 @@ export const setSearchResultCount = (resultCount: number): void => {
       }
       noResultsTimeout = setTimeout(
         triggerNoSearchResults.bind(null, s.eVar26),
-        1000,
+        1000
       );
     }
   }
 };
 
-export const trackSearchQuery = (
-  _input,
-  _event,
-  suggestion,
-  _datasetNumber,
-  _context,
-): void => {
-  if (typeof window !== "undefined" && typeof s != "undefined") {
+export const trackSearchQuery = (_input, _event, suggestion): void => {
+  if (typeof window !== 'undefined' && typeof s != 'undefined') {
     s.linkTrackVars =
-      "prop39,prop41,prop50,prop61,prop62,eVar26,eVar27,eVar39,eVar41,eVar50,eVar61,eVar62,eVar69,events";
-    s.linkTrackEvents = "event1";
-    s.events = "event1";
-    s.tl(true, "o", "internal search");
+      'prop39,prop41,prop50,prop61,prop62,eVar26,eVar27,eVar39,eVar41,eVar50,eVar61,eVar62,eVar69,events';
+    s.linkTrackEvents = 'event1';
+    s.events = 'event1';
+    s.tl(true, 'o', 'internal search');
   }
   window.location.assign(suggestion.url);
 };
@@ -157,12 +108,29 @@ export const trackFeedbackSubmission = (feedback: boolean) => {
       type: 'click',
       name: feedback ? 'YesVote' : 'NoVote'
     }
-  }
+  };
 
   AWSMA.ready(() => {
-    document.dispatchEvent(new CustomEvent(AWSMA.TRIGGER_EVENT, {detail: opt}));
+    document.dispatchEvent(
+      new CustomEvent(AWSMA.TRIGGER_EVENT, { detail: opt })
+    );
   });
-}
+};
+
+export const trackExpanderOpen = (expanderId) => {
+  const opt = {
+    event: {
+      type: 'click',
+      name: `ExpanderOpen.${expanderId}`
+    }
+  };
+
+  AWSMA.ready(() => {
+    document.dispatchEvent(
+      new CustomEvent(AWSMA.TRIGGER_EVENT, { detail: opt })
+    );
+  });
+};
 
 export const trackCopyClicks = (data) => {
   const opt = {
@@ -171,9 +139,43 @@ export const trackCopyClicks = (data) => {
       name: 'CopyCode'
     },
     data: data
-  }
+  };
 
   AWSMA.ready(() => {
-    document.dispatchEvent(new CustomEvent(AWSMA.TRIGGER_EVENT, {detail: opt}));
+    document.dispatchEvent(
+      new CustomEvent(AWSMA.TRIGGER_EVENT, { detail: opt })
+    );
   });
-}
+};
+
+// Track the click on the "Whats new" banner component
+export const trackWhatsNewBanner = () => {
+  const opt = {
+    event: {
+      type: 'click',
+      name: 'WhatsNewBanner'
+    }
+  };
+
+  AWSMA.ready(() => {
+    document.dispatchEvent(
+      new CustomEvent(AWSMA.TRIGGER_EVENT, { detail: opt })
+    );
+  });
+};
+
+// Track the click on the Version Switcher component
+export const trackVersionChange = (viewOld: boolean) => {
+  const opt = {
+    event: {
+      type: 'click',
+      name: `VersionChanged${viewOld ? 'Prev' : 'Current'}`
+    }
+  };
+
+  AWSMA.ready(() => {
+    document.dispatchEvent(
+      new CustomEvent(AWSMA.TRIGGER_EVENT, { detail: opt })
+    );
+  });
+};
