@@ -1,3 +1,3 @@
-import {join} from "path";
+import { join } from 'path';
 const [, , task] = process.argv;
 import(join(__dirname, task));
