@@ -2,6 +2,7 @@ const puppeteer = require('puppeteer');
 const axios = require('axios');
 
 const SITEMAP_URL = 'https://docs.amplify.aws/sitemap.xml';
+const DOMAIN = 'https://docs.amplify.aws';
 const CRAWLER_EXCEPTIONS = [
   'https://aaaaaaaaaa.execute-api.us-east-1.amazonaws.com/api',
   'https://aaaaaaaaaaaaaaaaaaaaaaaaaa.appsync-api.us-east-1.amazonaws.com/graphql',
@@ -11,12 +12,13 @@ const GITHUB_CREATE_ISSUE_LINK =
   'https://github.com/aws-amplify/docs/issues/new';
 const GITHUB_EDIT_LINK = 'https://github.com/aws-amplify/docs/edit/';
 
-const getSitemapUrls = async () => {
-  let browser = await puppeteer.launch();
+const getSitemapUrls = async (localDomain) => {
+  let browser = await puppeteer.launch({ headless: 'new' });
 
   const page = await browser.newPage();
 
-  let response = await page.goto(SITEMAP_URL);
+  let siteMap = localDomain ? `${localDomain}/sitemap.xml` : SITEMAP_URL;
+  let response = await page.goto(siteMap);
 
   const siteMapUrls = [];
 
@@ -33,6 +35,10 @@ const getSitemapUrls = async () => {
         urlTags,
         i
       );
+      if (localDomain) {
+        // Currently the sitemap is always generated with the prod docs domain so we need to replace this with localhost
+        url = url.replace(DOMAIN, localDomain);
+      }
       siteMapUrls.push(url);
     }
   }
@@ -42,42 +48,54 @@ const getSitemapUrls = async () => {
   return siteMapUrls;
 };
 
-const retrieveLinks = async (siteMapUrls, visitedLinks) => {
-  let browser = await puppeteer.launch();
+const retrieveLinks = async (siteMapUrls, visitedLinks, localDomain) => {
+  let browser = await puppeteer.launch({ headless: 'new' });
 
-  const page = await browser.newPage();
+  let page = await browser.newPage();
 
   const urlsToVisit = [];
 
   for (let i = 0; i < siteMapUrls.length; i++) {
     let url = siteMapUrls[i];
 
-    let response = await page.goto(url);
-    if (response && response.status() && response.status() === 200) {
-      visitedLinks[url] = true;
+    try {
+      let response = await page.goto(url, { waitUntil: 'domcontentloaded' });
+      await new Promise((r) => setTimeout(r, 100)); // localhost hangs on wait for idle so use a short timeout instead
+      if (response && response.status() && response.status() === 200) {
+        console.log(`successfully visited ${url} to retrieve links`);
+        visitedLinks[url] = true;
 
-      const urlList = await page.evaluate(async (url) => {
-        let urls = [];
-        let elements = document.getElementsByTagName('a');
-        for (let i = 0; i < elements.length; i++) {
-          let element = elements[i];
-          if (element.href) {
-            const link = {
-              url: element.href,
-              parentUrl: url,
-              linkText: element.textContent
-            };
-            urls.push(link);
+        const urlList = await page.evaluate(async (url) => {
+          let urls = [];
+          let elements = document.getElementsByTagName('a');
+          for (let i = 0; i < elements.length; i++) {
+            let element = elements[i];
+            if (element.href) {
+              const link = {
+                url: element.href,
+                parentUrl: url,
+                linkText: element.textContent
+              };
+              urls.push(link);
+            }
           }
-        }
-        return urls;
-      }, url);
+          return urls;
+        }, url);
 
-      urlList.forEach((link) => {
-        if (!CRAWLER_EXCEPTIONS.includes(link.url)) {
-          urlsToVisit.push(link);
-        }
-      });
+        urlList.forEach((link) => {
+          if (
+            !CRAWLER_EXCEPTIONS.includes(link.url) &&
+            (!localDomain || link.url.startsWith(localDomain))
+          ) {
+            urlsToVisit.push(link);
+          }
+        });
+      }
+    } catch (e) {
+      console.log(`failed to load ${url}: ${e}`);
+      browser.close();
+      browser = await puppeteer.launch({ headless: 'new' });
+      page = await browser.newPage();
     }
   }
 
@@ -97,14 +115,18 @@ const formatString = (inputs) => {
   return retString;
 };
 
-const linkChecker = async () => {
+const linkChecker = async (localDomain) => {
   const visitedLinks = {};
   const statusCodes = {};
   const brokenLinks = [];
 
-  const siteMapUrls = await getSitemapUrls();
+  const siteMapUrls = await getSitemapUrls(localDomain);
 
-  const urlsToVisit = await retrieveLinks(siteMapUrls, visitedLinks);
+  const urlsToVisit = await retrieveLinks(
+    siteMapUrls,
+    visitedLinks,
+    localDomain
+  );
 
   let allPromises = [];
 
@@ -120,7 +142,9 @@ const linkChecker = async () => {
     visitedLinks[href] = true;
 
     let request = axios
-      .get(href)
+      .get(href, {
+        timeout: 5000
+      })
       .then((response) => {
         let statusCode = response.status;
         if (statusCode && statusCode !== 200) {
@@ -135,12 +159,7 @@ const linkChecker = async () => {
           statusCodes[statusCode].push(href);
         }
         if (statusCode === 404) {
-          // this regular expression is meant to filter out any of the platform selector pages.  These are appearing in the result set
-          // because the crawler is seeing disabled platform dropdown links
-          const platformPages = /\/q\/(platform|integration|framework)\/(android|ios|flutter|js|react-native)/gm;
-          if (!platformPages.test(link.url)) {
-            brokenLinks.push(link);
-          }
+          brokenLinks.push(link);
         }
       });
 
@@ -156,7 +175,10 @@ const linkChecker = async () => {
 };
 
 module.exports = {
-  checkLinks: async () => {
+  checkProdLinks: async () => {
     return await linkChecker();
+  },
+  checkDevLinks: async () => {
+    return await linkChecker('http://localhost:3000');
   }
 };
